Added heimdal-dev as alternative for libkrb5-dev
The package builds fine without it. It does not have the functionality of krb5_get_error_message, but the freeradius code is already able to work around that limitation (using HAVE_KRB5_GET_ERROR_MESSAGE).

The main reason for this change is that the packages libkrb5-dev and heimdal-dev cannot both be installed on a machine, and Samba has a requirement on the heimdal-dev package. With this patch, my machine can happily compile Samba and FreeRADIUS.